task-a/flask/requirements.txt
flask
google-cloud-storage
google-cloud-firestore
gunicorn